FLIZ-432 refactor: NotificationAccordion 및 NotificationSideBar 파일 위치 변경
