+++ apps/trainer/app/notification/reservation/page.tsx
@@ -9,11 +9,11 @@ import { ToggleGroup, ToggleGroupItem } from "@ui/components/ToggleGroup";
 import { useRouter } from "next/navigation";
 import { Fragment, Suspense, useEffect, useRef, useState } from "react";
 
+import NotificationSideBar from "@trainer/app/notification/_components/NotificationSideBar";
 import { notificationQueries } from "@trainer/queries/notification";
 import { useNotificationStore } from "@trainer/store/notificationStore";
 
 import Logo from "@trainer/components/Logo";
-import NotificationSideBar from "@trainer/components/NotificationSideBar";
 
 import useIntersectionObserver from "@trainer/hooks/useIntersectionObserver";
 
@@ -118,7 +118,7 @@ function ReservationNotificationContent({
         <Header.Left>
           <NotificationSideBar />
         </Header.Left>
-        <Header.Title content="예약 요청" />
+        <Header.Title content={NOTIFICATION_TYPE} />
         <Header.Right>
           <NotificationSearch />
         </Header.Right>
